create fig2, updates to fig1 and table1

+++ LaTeX_documentation/SciData_manuscript/Figure_Table_Scripts/Map_Monitor_Locations.R
@@ -34,17 +34,19 @@ title_string <- "Observation Locations"
 caption_string <- "Locations of PM\\textsubscript{2.5} monitoring stations that had at least one observation during our study period (2008-2018)."
 ClearPage = FALSE
 LatexFileName <- file.path("/home/rstudio","estimate-pm25","LaTeX_documentation","SciData_manuscript","Fig1MonitorLocations.tex")
-if (exists(LatexFileName)) {
-  file.remove(LatexFileName) # delete file if an old version is there
-}
+#if (exists(LatexFileName)) {
+#  file.remove(LatexFileName) # delete file if an old version is there
+#}
 
+if (file.exists(LatexFileName) == TRUE) {file.remove(LatexFileName)}
+image_file_type = "pdf"
 
-  FigFileName <- Plot_to_ImageFile_TopOnly.fn(output.directory, file_sub_label, plot_name_extension = plot_name_extension) # start image file
+  FigFileName <- Plot_to_ImageFile_TopOnly.fn(output.directory, file_sub_label, plot_name_extension = plot_name_extension,image_file_type) # start image file
   WestCountymapGeom <- map_county_base_layer.fn(CountyMaps.directory, study_states_abbrev) # create map of counties
   plot_color <- "blue"
   points(this_df$Longitude,this_df$Latitude,pch = 19,col=plot_color)#, 
   Plot_to_ImageFile_BottomOnly.fn(FigFileName = FigFileName, title_string = title_string) # finish image file
-  LaTex_code_4_figure.fn(LatexFileName = LatexFileName, title_string = caption_string, file_sub_label = file_sub_label, plot_name_extension = plot_name_extension, output.directory.short = output.directory.short, image_format = "jpg", ClearPage = ClearPage)
+  LaTex_code_4_figure.fn(LatexFileName = LatexFileName, title_string = caption_string, file_sub_label = file_sub_label, plot_name_extension = plot_name_extension, output.directory.short = output.directory.short, image_format = image_file_type, ClearPage = ClearPage)
   
 #### Code below this line probably obsolete ####
   #LaTex_code_4_figure.fn(LatexFileName = LatexFileName, title_string = title_string, file_sub_label = file_sub_label, plot_name_extension = plot_name_extension, output.directory.short = output.directory.short, image_format = "jpg", ClearPage = ClearPage)
